+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 ------------------
 
 - #2775 Auto-add CCContact when hidden on Sample Add form
+- #2831 Migrate ARReport to Dexterity
+- #2838 Fix 'add_permission' error when upgrading from 2711 to 2718
+- #2835 New Sidebar Navigation
 - #2834 Handle tat as timedelta or dict during AT to DX migration
 - #2832 Fix sample type prefix validation
 - #2837 Disable Cancel Button in Sample Registration Form after Save